Enhancement: Renamed exceptions folder into meaningful text
